% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-fingerprints.R
\name{download_fingerprints}
\alias{download_fingerprints}
\title{Download fingerprints from the Recog repo}
\usage{
download_fingerprints(download_path)
}
\arguments{
\item{download_path}{path to a directory where the most current Recog
fingerprints will be downloaded to and compressed.}
}
\description{
Download fingerprints from the Recog repo
}
\note{
This uses \code{GITHUB_PAT} from the environment, if available
}
