% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_fingerprints.R
\name{print.recog_fingerprints}
\alias{print.recog_fingerprints}
\alias{print.recog_fingerprint}
\title{Print fingerprints}
\usage{
\method{print}{recog_fingerprints}(x, ...)

\method{print}{recog_fingerprint}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{unused}

\item{x}{object}

\item{...}{unused}
}
\description{
Print fingerprints

Print a fingerprint
}
\keyword{internal}
