% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-fingerprints.R
\name{load_fingerprints}
\alias{load_fingerprints}
\title{Load a directory of fingerprints}
\usage{
load_fingerprints(fingerprints_directory)
}
\arguments{
\item{fingerprints_directory}{a directory where Recogs XML fingerprints are stored}
}
\value{
a \code{list} of \code{recog_fingerprints}
}
\description{
Load a directory of fingerprints
}
\note{
You \strong{cannot} serialize the resultant object as it uses the \code{ore}
regex library and those compiled patterns are not serializable.
}
\examples{
load_fingerprints(
  system.file("fingerprints", package="rrecog")
)
}
