% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_fingerprints.R
\name{read_fingerprints_file}
\alias{read_fingerprints_file}
\title{Ingest Recog XML fingerprints from a file and precompile regular expressions}
\usage{
read_fingerprints_file(path)
}
\arguments{
\item{path}{path to a Recog XML fingerprints file}
}
\value{
a \code{recog_fingerprints} object which is a list containing \code{recog_fingerprint} elements
}
\description{
Ingest Recog XML fingerprints from a file and precompile regular expressions
}
\examples{
read_fingerprints_file(
  system.file("fingerprints", "http_servers.xml.gz", package="rrecog")
)
}
