% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrecog-package.R
\docType{package}
\name{rrecog}
\alias{rrecog}
\alias{rrecog-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/rrecog}
\item BugReports: \url{https://gitlab.com/hrbrmstr/rrecog/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
