% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-fingerprints.R
\name{use_builtin_fingerprints}
\alias{use_builtin_fingerprints}
\title{Use built-in fingerprints}
\usage{
use_builtin_fingerprints()
}
\description{
Use a snapshot of the Recog repository was taken on 2018-08-17. There may be
newer fingerprints or updates to existing ones in the repository. Use
\code{\link[=download_fingerprints]{download_fingerprints()}} to download, compress and store new fingerprints
from the Recog repository.
}
\note{
this is simply \code{load_fingerprints(system.file("fingerprints", package="rrecog"))}

You \strong{cannot} serialize the resultant object as it uses the \code{ore}
regex library and those compiled patterns are not serializable.
}
\examples{
use_builtin_fingerprints()
}
