% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-matches.R
\name{recog_match}
\alias{recog_match}
\title{Find fingerprint matches for a given source}
\usage{
recog_match(recog_db, source, protocols = ".*", matches = NULL)
}
\arguments{
\item{recog_db}{a structure created with \code{\link[=load_fingerprints]{load_fingerprints()}} or
\code{\link[=use_builtin_fingerprints]{use_builtin_fingerprints()}}}

\item{source}{the 1-element character vector to compare against}

\item{protocol, matches}{regexs to limit what you're comparing against. These
are boolean \strong{OR'd} together}
}
\description{
This is an exhaustive lookup for the fingerprint in all the
\code{protocol}/\code{matches} categories. As a result, it's not very fast
on its own. However, the function has been \code{\link[memoise:memosie]{memoise::memosie()}}'d.
As such, if you are performing a number of recogs in a single
R session and working from a typical data source (i.e. a large file
with many common strings, such as a collection of HTTP `Server``
header strings), you will see performance gains after each distinct
match input.
}
\examples{
recog_db <- use_builtin_fingerprints()
recog_match(recog_db, "VShell_Special_Edition_2_5_0_204 VShell", "ssh")
}
